UI追加による変更

+++ Game/App/Library/UserInterface.h
@@ -13,7 +13,7 @@ namespace My3dApp
     public:
         static void CreateInstance();
         static void DeleteInstance();
-        static void UIText(int x, int y, int color, const char* str, int num);
+        static void UIText(int x, int y, int color, const char* str, float num);
         static void UIText(int x, int y, int color, const char* str);
         static void UIBox(int x1, int x2, int y1, int y2, int frameSize, int mainCr, int frameCr);
     };
